+++ src/modules/testModule.js
@@ -1,0 +1,4 @@
+import { unknownObject, createMyServer } from './cjsToEsm.mjs';
+console.log(unknownObject);
+createMyServer.listen(3000);
+console.log('Server started on port 3000');
